move storage to lib
