+++ pombot/lib/storage.py
